% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{db_last_updates}
\alias{db_last_updates}
\title{List providers and datasets sorted by creation date}
\usage{
db_last_updates()
}
\description{
List providers and datasets sorted by creation date
}
\examples{
\donttest{
db_last_updates()
}
}
