% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_providers}
\alias{db_providers}
\title{List available providers}
\usage{
db_providers(provider_code = NULL)
}
\arguments{
\item{provider_code}{\code{character(1)} code of the provider}
}
\value{
A \code{data.frame()} with the available providers.
}
\description{
List available providers
}
\examples{
\donttest{
db_providers()
db_providers("BEA")
}
}
