% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_search}
\alias{db_search}
\title{List datasets from full-text search}
\usage{
db_search(query)
}
\arguments{
\item{query}{\code{character(1)} full-text search.}
}
\description{
List datasets from full-text search
}
\examples{
\donttest{
db_search("GDP France")
}
}
