% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_series}
\alias{db_series}
\title{List series}
\usage{
db_series(
  series_ids,
  facets = FALSE,
  observations = FALSE,
  metadata = TRUE,
  format = "json",
  align_periods = FALSE
)
}
\description{
List series
}
\examples{
\donttest{
db_series("AMECO/ZUTN/EA19.1.0.0.0.ZUTN")
}
}
