% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnomics-package.R
\docType{package}
\name{dbnomics-package}
\alias{dbnomics}
\alias{dbnomics-package}
\title{dbnomics: Client for dbnomics}
\description{
Download and search data from dbnomics
}
\seealso{
Useful links:
\itemize{
  \item \url{https://m-muecke.github.io/dbnomics/}
  \item \url{https://github.com/m-muecke/dbnomics}
  \item Report bugs at \url{https://github.com/m-muecke/dbnomics/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Mücke \email{muecke.maximilian@gmail.com} (\href{https://orcid.org/0009-0000-9432-9795}{ORCID})

}
\keyword{internal}
