% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnomics.R
\name{get_dbnomics}
\alias{get_dbnomics}
\title{Get dbnomics}
\usage{
get_dbnomics(endpoint, slug, suffix)
}
\arguments{
\item{endpoint}{a string either "providers", "datasets" or "series"}

\item{slug}{a string name of the slug}

\item{suffix}{a suffix. For instance "values"}
}
\value{
a get request
}
\description{
Get dbnomics
}
\examples{

get_dbnomics(endpoint = "series", slug = "bis-pp-ls-q-au")


# https://api.db.nomics.world/api/v1/json/datasets/insee-cna-2010-dep-apu/values
get_dbnomics(
endpoint = "datasets",
slug = "insee-cna-2010-dep-apu",
suffix = "values")

\dontrun{
library("magrittr")
get_dbnomics(
endpoint = "datasets",
slug = "insee-cna-2010-dep-apu",
suffix = "values") \%>\%
httr::content() \%>\%
magrittr::extract2("data")
}

}
